Modify uninstall-homebrew.sh to work on M1 Macs

+++ tests/uninstall-homebrew.sh
@@ -8,6 +8,19 @@ chmod +x ./uninstall.sh
 sudo ./uninstall.sh --force
 
 # Clean up Homebrew directories.
-sudo rm -rf /usr/local/Homebrew
-sudo rm -rf /usr/local/Caskroom
-sudo rm -rf /usr/local/bin/brew
+
+# For pre m1 macs:
+if [[ -d /usr/local/Homebrew ]]; then
+    sudo rm -rf /usr/local/Homebrew
+fi
+if [[ -d /usr/local/Caskroom ]]; then
+    sudo rm -rf /usr/local/Caskroom
+fi
+if [[ -f /usr/local/bin/brew ]]; then
+    sudo rm -rf /usr/local/bin/brew
+fi
+
+# For post m1 macs:
+if [[ -d /opt/homebrew ]]; then
+    sudo rm -rf /opt/homebrew
+fi
